fix: networks serialisation
Deserialization was attempting to deserialize via snake case, however the `json` definitions in status-go are using camel case.

The odd thing here is that when *serializing*, status-go expects the fields to be snake cased. Ideally, these should be consistent so that we can use both `Json.encode` and `Json.decode`.