fix(@desktop/chat): replies are not working in communities
This check is here, because of issue#3490, that was happening due to different
messages' order (received from status go). We were emitting always a message on
index 0, what in some cases was not the reply message (doesn't contain "responseTo"
value). This should be handled in status-go but since that part is used by mobile
also and it is working for them, change is applied here.

Fixes: #3490