chore: add GitHub Actions workflow to run tests
Closes #65.

There's still some work to be done. With the changes in this PR, this project's
Makefile is a bit of a hybrid between the approaches in the Makefiles of
status-desktop and nim-status. In the future, as this project's tests expand in
scope, the logic in the Makefile, `nim_status_lib.nimble`, and/or
`.github/workflows/test.yml` may need to be revised.